FB test

+++ js/controllers/offerDetailsCtrl.js
@@ -11,47 +11,66 @@ var offerDetailsCtrl = (function () {
 		query.find({
 			success: function (offer) {
 				var offerDetailsTemplate = $('#offer-details').html();
-				var container = $('#offer-container');
+				var container = $('.panel-body');
 
 				var outputOfferHtml = Mustache.render(offerDetailsTemplate, offer[0]._serverData);
 				container.append(outputOfferHtml);
-				
+
 				renderSellerInfo(offer[0]._serverData.createdBy.id);
+				changeFbShareButtonURL();
 				addDeleteButton(offer);
 			}
 		})
 	}
 
 	function addDeleteButton(offer) {
-		if (offer[0]._serverData.createdBy.id === Parse.User.current().id) {
-			var deleteBtn = $('<button/>').addClass('btn btn-danger pull-right btn-lg').html('Delete offer');
+		if (Parse.User.current()) {
+			if (offer[0]._serverData.createdBy.id === Parse.User.current().id) {
+				var deleteBtn = $('<button/>').addClass('btn btn-danger pull-right btn-lg').html('Delete offer');
 
-			deleteBtn.on('click', function () {
-				offer[0].destroy({
-					success: function (offer) {
-						window.history.back();
-					}
-				});
-			})
+				deleteBtn.on('click', function () {
+					offer[0].destroy({
+						success: function (offer) {
+							window.history.back();
+						}
+					});
+				})
 
-			$('.panel-body').append(deleteBtn);
+				$('.panel-body').append(deleteBtn);
+			}
 		}
 	}
 
 	function renderSellerInfo(sellerId) {
 		var User = Parse.Object.extend('User');
 		var query = new Parse.Query(User);
-		
+
 		query.equalTo('objectId', sellerId);
 		query.find({
 			success: function (seller) {
 				var sellerDetailsTemplate = $('#seller-details').html();
 				var container = $('#seller-info');
-				
+
 				var outputSellerHtml = Mustache.render(sellerDetailsTemplate, seller[0]._serverData);
 				container.append(outputSellerHtml);
 			}
 		})
+	}
+
+	function changeFbShareButtonURL() {
+		$('#share').on('click', function () {
+			
+			FB.ui({
+				method: 'share',
+				href: 'http://127.0.0.1:8080/#/offerDetails/:puIu1P2PHT'
+			}, function(response){});
+		})
+		
+		console.log(window.location.href);
+		var button = $('.fb-share-button')[0];
+		button.baseURI = window.location.href;
+		console.log(button.baseURI);
+		console.log($('.fb-share-button')[0].baseURI)
 	}
 
 	return {
